% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hyper_randomwalk.R
\name{get_hyper_randomwalk}
\alias{get_hyper_randomwalk}
\title{get_hyper_randomwalk}
\usage{
get_hyper_randomwalk(P, H, theta = 0.8)
}
\arguments{
\item{P}{the transition possibility matrix}

\item{H}{adjacent matrix for i-th sample}

\item{theta}{the restart probability at every step of the random walk}
}
\description{
get_hyper_randomwalk uses the transition possibility matrix, the mutation data to perform random walk with restart on hypergraph
theta is the restart probability at every step of the random walk.
output of get_hyper_randomwalk is a list which combines the Importance of the genes and the distance of vectors between one iteration
}
