% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mut_data.R
\name{get_mut_data}
\alias{get_mut_data}
\title{get_mut_data}
\usage{
get_mut_data(mutation_file)
}
\arguments{
\item{mutation_file}{the text that contains the mutation data, a gene*sample matrix with rownames being gene names and column names being sample IDs}
}
\value{
Mutation, the mutation data, rows represent genes and columns represent samples
}
\description{
get_mut_data uses reads the mutation file
it deletes some row(s) of the mutation data whose rowsum is zero, rownames of the remaining rows are considered possibale genes
}
\examples{


}
