% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_P.R
\name{get_P}
\alias{get_P}
\title{get_P}
\usage{
get_P(H, vertex_W, hyperedge_W)
}
\arguments{
\item{H}{adjacent matrix for i-th sample}

\item{vertex_W}{the vertex weight of the hypergraph}

\item{Mutation}{the mutation data, rows represent genes and columns represent samples}
}
\description{
get_P uses the mutation data, expression data and the vertex weight of the hypergraph to calculate the transition possibility matrix
output of get_P is the transition possibility matrix of the hypergraph
}
\examples{

}
