% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vertex_W.R
\name{get_vertex_W}
\alias{get_vertex_W}
\title{get_vertex_W}
\usage{
get_vertex_W(
  sample_i,
  data_mut_com,
  com_mut_samples,
  Graph,
  sample_i_mut_genes,
  H
)
}
\arguments{
\item{sample_i}{the i-th sample}

\item{data_mut_com}{usable mutation data}

\item{com_mut_samples}{other samples that have co-mutated gene(s) with the i-th sample}

\item{Graph}{the graph object in igraph format}

\item{sample_i_mut_genes}{placeholder}

\item{H}{adjacent matrix for i-th sample}
}
\description{
get_vertex_W uses the mutation data and the graph constructed from network data to calculate the weight of the hypergraph
output of get_vertex_W is the vertex weight of the hypergraph
}
\examples{


}
