% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hygraph_randomwalk.R
\name{hygraph_randomwalk}
\alias{hygraph_randomwalk}
\title{hygraph_randomwalk}
\usage{
hygraph_randomwalk(sample_i, data_mut_com, co_mut, correlation_samples, Graph)
}
\arguments{
\item{sample_i}{the i-th sample}

\item{data_mut_com}{usable mutation data}

\item{correlation_samples}{similarity matrix for the samples}

\item{Graph}{the graph object in igraph format}

\item{com_mut}{other samples that have co-mutated gene(s) with the i-th sample}
}
\value{
importance score for i-th sample
}
\description{
hygraph_randomwalk calculates the importance score for each sample using pre-processed data
}
\examples{


}
