\name{STRINGv10}
\alias{STRINGv10}
\docType{data}
\title{
STRINGv10
}
\description{
The STRING database (http://string-db.org) aims to provide a critical assessment and integration of protein-protein interactions, including direct (physical) as well as indirect (functional) associations. The new version 10.0 of STRING covers more than 2000 organisms.
}
\usage{data(STRINGv10)}
\details{
Protein-Protein Interaction Networks.
}
\source{
https://string-db.org
}
\references{
D. Szklarczyk, A. Franceschini, S. Wyder, K. Forslund, D. Heller,
J. Huerta-Cepas, M. Simonovic, A. Roth, A. Santos, K. P. Tsafou,
Nucleic Acids Res. 2014, 43, D447.
}
\examples{
data(STRINGv10)
}
\keyword{datasets}
